Fix GasGraph construction, remove warnings and other details

+++ project/src/entity/core/impl/AdjacencyListGraph.cpp
@@ -10,7 +10,7 @@ AdjacencyListGraph::AdjacencyListGraph(int vertexAmount) {
 
 void AdjacencyListGraph::resetSize(int newVerticesSize) {
     this->vertexAdjacents = std::vector<std::list<AdjacencyNode*>>(newVerticesSize);
-    for(int index = 0; index < this->vertexAdjacents.size(); index++){
+    for(size_t index = 0; index < this->vertexAdjacents.size(); index++){
         this->vertexAdjacents.at(index) = std::list<AdjacencyNode*>();
     }
 }
